- Fix minor compile warnings.
git-svn-id: https://anonsvn.ncbi.nlm.nih.gov/repos/v1/trunk/c++@21567 78c7ea69-d796-4a43-9a09-de51944f1b03

+++ src/objtools/readers/seqdb/seqdbcommon.cpp
@@ -329,7 +329,7 @@ bool CSeqDBGiList::GiToOid(int gi, int & oid, int & index)
 {
     InsureOrder(eGi);  // would assert be better?
     
-    int b(0), e(m_GisOids.size());
+    int b(0), e((int)m_GisOids.size());
     
     while(b < e) {
         int m = (b + e)/2;
